feat: update workspace crates, add API tests, and improve project infrastructure
- Update dependency versions across all workspace crates (Cargo.lock +466)
- Add resolver v3, workspace lints, and coverage profile to root Cargo.toml
- Add servers/api integration tests (containers, tracing, E2E, UI)
- Add e2e agent test for frontend
- Add milestone audit doc and test plan
- Add test runner scripts (scripts/test/)
- Update admin service, sync engine, and agent route implementations
- Regenerate contract bindings (AdminDashboardStats, ToolCall)
- Add nextest config, update clippy.toml

+++ packages/adapters/storage/libsql/src/embedded.rs
@@ -75,7 +75,7 @@ impl LibSqlPort for EmbeddedLibSql {
         while let Some(row) = rows.next().await? {
             let mut map = serde_json::Map::new();
             for i in 0..row.column_count() {
-                let name = row.column_name(i as i32).unwrap_or_default();
+                let name = row.column_name(i).unwrap_or_default();
                 let value = match row.get_value(i)? {
                     libsql::Value::Null => serde_json::Value::Null,
                     libsql::Value::Integer(n) => serde_json::json!(n),
